"Adding PropertyController endpoints and removing email service"
