--- portal/src/main/java/com/propertymanagement/portal/EmailService/EmailService.java
@@ -1,8 +1,0 @@
-package com.propertymanagement.portal.EmailService;
-
-import java.io.IOException;
-
-public interface EmailService {
-
-    public void sendEmail(String to_address, String subject, String body) throws IOException;
-}
