+++ portal/src/main/java/com/propertymanagement/portal/PortalApplication.java
@@ -28,7 +28,7 @@ public class PortalApplication {
 		SpringApplication.run(PortalApplication.class, args);
 	}
 
-	@Bean
+	/*@Bean
 	public CommandLineRunner commandLineRunner(
 			AuthenticationService service
 	) {
@@ -51,7 +51,7 @@ public class PortalApplication {
 					.build();
 			System.out.println("Customer token: " + service.registerCustomer(customer).getAccessToken());
 		};
-	}
+	}*/
 	@Bean
 	public ModelMapper modelMapper() {
 		return new ModelMapper();
