+++ portal/src/main/java/com/propertymanagement/portal/controller/PropertyController.java
@@ -1,13 +1,17 @@
 package com.propertymanagement.portal.controller;
 
 import com.propertymanagement.portal.domain.ListingType;
+import com.propertymanagement.portal.domain.Offer;
 import com.propertymanagement.portal.domain.Property;
 import com.propertymanagement.portal.domain.PropertyType;
+import com.propertymanagement.portal.dto.request.MakeOfferRequest;
 import com.propertymanagement.portal.service.PropertyService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
+import org.springframework.http.HttpStatus;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -50,6 +54,55 @@ public class PropertyController {
         Pageable pageable = PageRequest.of(page, size);
         return propertyService.findPropertiesByCriteria(minPrice, maxPrice, listingType, minBedRooms, maxBedRooms,
                 minBathRooms, maxBathRooms, propertyType, city, state, pageable);
+    }
+
+    @PreAuthorize("hasAuthority('CUSTOMER')")
+    @ResponseStatus(HttpStatus.CREATED)
+    @PostMapping("/{propertyId}/offers")
+    public Offer makeOffer(@PathVariable Long propertyId, @RequestBody MakeOfferRequest makeOfferRequest){
+        return propertyService.makeOffer(propertyId, makeOfferRequest);
+    }
+
+    @PreAuthorize("hasAuthority('CUSTOMER')")
+    @ResponseStatus(HttpStatus.OK)
+    @PostMapping("/{propertyId}/like")
+    public void addToFavourites(@PathVariable Long propertyId){
+        propertyService.addToFavourites(propertyId);
+    }
+
+    @PreAuthorize("hasAuthority('CUSTOMER')")
+    @ResponseStatus(HttpStatus.OK)
+    @PostMapping("/{propertyId}/unlike")
+    public void removeFromFavourites(@PathVariable Long propertyId){
+        propertyService.removeFromFavourites(propertyId);
+    }
+
+    @PreAuthorize("hasAuthority('CUSTOMER')")
+    @ResponseStatus(HttpStatus.OK)
+    @PutMapping("/{propertyId}/offers")
+    public Offer updateOffer(@PathVariable Long propertyId, @RequestBody MakeOfferRequest makeOfferRequest){
+        return propertyService.updateOffer(propertyId, makeOfferRequest);
+    }
+
+    @PreAuthorize("hasAuthority('CUSTOMER')")
+    @ResponseStatus(HttpStatus.OK)
+    @DeleteMapping("/{propertyId}/offers")
+    public void deleteOffer(@PathVariable Long propertyId){
+        propertyService.deleteOffer(propertyId);
+    }
+
+    @PreAuthorize("hasAuthority('OWNER')")
+    @ResponseStatus(HttpStatus.OK)
+    @PutMapping("/{propertyId}/offers/{offerId}/accept")
+    public void acceptOffer(@PathVariable Long propertyId, @PathVariable Long offerId){
+        propertyService.acceptOffer(propertyId, offerId);
+    }
+
+    @PreAuthorize("hasAuthority('OWNER')")
+    @ResponseStatus(HttpStatus.OK)
+    @PutMapping("/{propertyId}/offers/{offerId}/reject")
+    public void rejectOffer(@PathVariable Long propertyId, @PathVariable Long offerId){
+        propertyService.rejectOffer(propertyId, offerId);
     }
