+++ portal/src/main/java/com/propertymanagement/portal/service/impl/PropertyServiceImpl.java
@@ -1,9 +1,10 @@
 package com.propertymanagement.portal.service.impl;
 
-import com.propertymanagement.portal.domain.ListingType;
-import com.propertymanagement.portal.domain.Property;
-import com.propertymanagement.portal.domain.PropertyType;
-import com.propertymanagement.portal.repository.PropertyRespository;
+import com.propertymanagement.portal.domain.*;
+import com.propertymanagement.portal.dto.request.MakeOfferRequest;
+import com.propertymanagement.portal.exception.InvalidInputException;
+import com.propertymanagement.portal.exception.RecordNotFoundException;
+import com.propertymanagement.portal.repository.*;
 import com.propertymanagement.portal.service.PropertyService;
 import jakarta.persistence.EntityManager;
 import jakarta.persistence.PersistenceContext;
@@ -15,8 +16,11 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
 import org.springframework.data.domain.Pageable;
+import org.springframework.security.core.Authentication;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -24,6 +28,16 @@ import java.util.List;
 public class PropertyServiceImpl implements PropertyService {
     @Autowired
     PropertyRespository propertyRespository;
+    @Autowired
+    private CustomerRepository customerRepository;
+    @Autowired
+    private PropertyRepository propertyRepository;
+
+    @Autowired
+    private OfferRepository offerRepository;
+
+    @Autowired
+    private OwnerRepository ownerRepository;
 
     @PersistenceContext
     private EntityManager entityManager;
@@ -113,5 +127,124 @@ public class PropertyServiceImpl implements PropertyService {
     public List<Property> findAllProperties() {
         return propertyRespository.findAll();
     }
+
+
+    @Override
+    public Offer makeOffer(Long propertyId, MakeOfferRequest makeOfferRequest) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer customer = customerRepository.findCustomerByUserEmail(authentication.getName());
+        Property property = propertyRepository.findById(propertyId).orElseThrow(() -> new RecordNotFoundException("Property not found with id: " + propertyId));
+
+        if (property.getStatus()== PropertyStatus.CONTINGENT){
+            throw new InvalidInputException("Property is in Contingent status, you cannot make an offer for this property");
+        }
+        Offer offer = new Offer();
+        System.out.println("Customer: " + customer);
+        System.out.println("Authentication: " + authentication.getName());
+
+        //Check if the user already made an offer for the property
+        if (customer.getOffers().stream().anyMatch(o -> o.getProperty().getId().equals(propertyId))) {
+            throw new InvalidInputException("You have already made an offer for this property");
+        }
+
+        offer.setOfferAmount(makeOfferRequest.getOfferAmount());
+        offer.setOfferDate(LocalDateTime.now());
+        offer.setOfferStatus(OfferStatus.PENDING);
+        offer.setCustomer(customer);
+        offer.setProperty(property);
+        offer.setOfferType(makeOfferRequest.getOfferType());
+
+        customer.addOffer(offer);
+        property.addOffer(offer);
+        offerRepository.save(offer);
+        customerRepository.save(customer);
+        return offer;
+    }
+
+    @Override
+    public Offer updateOffer(Long propertyId, MakeOfferRequest makeOfferRequest) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer customer = customerRepository.findCustomerByUserEmail(authentication.getName());
+        Property property = propertyRepository.findById(propertyId).orElseThrow(() -> new RecordNotFoundException("Property not found with id: " + propertyId));
+        if (property.getStatus()== PropertyStatus.CONTINGENT){
+            throw new InvalidInputException("Property is in Contingent status, you cannot make an offer for this property");
+        }
+        Offer offer = offerRepository.findOfferByCustomerIdAndPropertyId(customer.getId(), propertyId);
+        //Check if the user already made an offer for the property
+        if (offer==null) {
+            throw new InvalidInputException("You have not made an offer for this property");
+        }
+
+        offer.setOfferAmount(makeOfferRequest.getOfferAmount());
+        offer.setOfferDate(LocalDateTime.now());
+        offer.setOfferStatus(OfferStatus.PENDING);
+        offer.setCustomer(customer);
+        offer.setProperty(property);
+        offer.setOfferType(makeOfferRequest.getOfferType());
+
+        customer.addOffer(offer);
+        property.addOffer(offer);
+        offerRepository.save(offer);
+        customerRepository.save(customer);
+        return offer;
+    }
+
+    @Override
+    public void deleteOffer(Long propertyId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer customer = customerRepository.findCustomerByUserEmail(authentication.getName());
+        Property property = propertyRepository.findById(propertyId).orElseThrow(() -> new RecordNotFoundException("Property not found with id: " + propertyId));
+        if (property.getStatus()== PropertyStatus.CONTINGENT){
+            property.setStatus(PropertyStatus.AVAILABLE);
+        }
+        Offer offer = offerRepository.findOfferByCustomerIdAndPropertyId(customer.getId(), propertyId);
+        //Check if the user already made an offer for the property
+        if (offer==null) {
+            throw new InvalidInputException("You have not made an offer for this property");
+        }
+        customer.removeOffer(offer);
+        property.removeOffer(offer);
+        offerRepository.delete(offer);
+    }
+
+    @Override
+    public void acceptOffer(Long propertyId, Long offerId) {
+        Property property = propertyRepository.findById(propertyId).orElseThrow(() -> new RecordNotFoundException("Property not found with id: " + propertyId));
+        Offer offer = offerRepository.findById(offerId).orElseThrow(() -> new RecordNotFoundException("Offer not found with id: " + offerId));
+        if (property.getStatus()== PropertyStatus.CONTINGENT){
+            throw new InvalidInputException("Property is in Contingent status, you cannot accept an offer for this property");
+        }
+        // Check if any other offer is accepted
+        if (property.getOffers().stream().anyMatch(o -> o.getOfferStatus().equals(OfferStatus.ACCEPTED))) {
+            throw new InvalidInputException("Another offer has already been accepted for this property");
+        }
+
+        offer.setOfferStatus(OfferStatus.ACCEPTED);
+        property.setStatus(PropertyStatus.PENDING);
+        propertyRepository.save(property);
+    }
+    @Override
+    public void rejectOffer(Long propertyId, Long offerId) {
+        /////////
+    }
+
+    @Override
+    public void addToFavourites(Long propertyId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer customer = customerRepository.findCustomerByUserEmail(authentication.getName());
+        Property property = propertyRepository.findById(propertyId).orElseThrow(() -> new RecordNotFoundException("Property not found with id: " + propertyId));
+        customer.addFavoutite(property);
+        customerRepository.save(customer);
+    }
+
+    @Override
+    public void removeFromFavourites(Long propertyId) {
+        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
+        Customer customer = customerRepository.findCustomerByUserEmail(authentication.getName());
+        Property property = propertyRepository.findById(propertyId).orElseThrow(() -> new RecordNotFoundException("Property not found with id: " + propertyId));
+        customer.removeFavourite(property);
+        customerRepository.save(customer);
+    }
+
 
 }
